+++ site/changelog.md
@@ -1,5 +1,15 @@
 # TorrentFile
 
+## Version 0.8.11
+
+- added new command line option for magnet subcommand
+- fixed some coverage gaps caused by 0.8.10
+- fixed bugs introduced by 0.8.10 with magnet URI creation
+- added support for hybrid magnet links
+- added tests for when a magnet link is generated with no trackers at all
+
+---
+
 ## Version 0.8.10
 
 - added support for bittorrent v2 magnet links